requirements.txt
fastapi
uvicorn
groq
langgraph
ddgs
python-dotenv
